feat: 공통 모듈을 common & mvc & reactive 모듈로 분리
