+++ src/backend/common-module/common-mvc/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    // WebMVC
+    implementation "org.springframework.boot:spring-boot-starter-web"
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+}
