+++ src/backend/common-module/common-mvc/src/main/java/com/lalala/mvc/config/CommonMvcModuleConfig.java
@@ -1,0 +1,21 @@
+package com.lalala.mvc.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.lalala.config.CommonModuleConfig;
+import com.lalala.mvc.passport.PassportExtractor;
+
+@Configuration
+public class CommonMvcModuleConfig extends CommonModuleConfig {
+
+    public CommonMvcModuleConfig(ObjectMapper objectMapper) {
+        super(objectMapper);
+    }
+
+    @Bean
+    public PassportExtractor passportExtractor() {
+        return new PassportExtractor(objectMapper, passportValidator());
+    }
+}
