+++ src/backend/common-module/common-mvc/src/main/java/com/lalala/mvc/exception/GlobalExceptionHandler.java
@@ -1,0 +1,113 @@
+package com.lalala.mvc.exception;
+
+import java.net.ConnectException;
+import java.nio.file.AccessDeniedException;
+
+import lombok.extern.slf4j.Slf4j;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindException;
+import org.springframework.web.HttpRequestMethodNotSupportedException;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
+
+import com.lalala.exception.BusinessException;
+import com.lalala.exception.ErrorCode;
+import com.lalala.exception.ErrorResponse;
+
+@ControllerAdvice
+@Slf4j
+public class GlobalExceptionHandler {
+    private ResponseEntity<ErrorResponse> handleException(ErrorResponse response) {
+        return new ResponseEntity<>(response, HttpStatus.valueOf(response.getStatus()));
+    }
+
+    @ExceptionHandler(ConnectException.class)
+    protected ResponseEntity<ErrorResponse> handleConnectException(
+            MethodArgumentNotValidException e) {
+        log.error("handleConnectException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.DATABASE_ERROR);
+        return handleException(response);
+    }
+
+    /**
+     * javax.validation.Valid or @Validated 으로 binding error 발생시 발생한다. HttpMessageConverter 에서 등록한
+     * HttpMessageConverter binding 못할경우 발생 주로 @RequestBody, @RequestPart 어노테이션에서 발생
+     */
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    protected ResponseEntity<ErrorResponse> handleMethodArgumentNotValidException(
+            MethodArgumentNotValidException e) {
+        log.error("handleMethodArgumentNotValidException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.INVALID_PARAMETER);
+        return handleException(response);
+    }
+
+    /**
+     * @ModelAttribut 으로 binding error 발생시 BindException 발생한다. ref
+     * https://docs.spring.io/spring/docs/current/spring-framework-reference/web.html#mvc-ann-modelattrib-method-args
+     */
+    @ExceptionHandler(BindException.class)
+    protected ResponseEntity<ErrorResponse> handleBindException(BindException e) {
+        log.error("handleBindException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.INVALID_PARAMETER);
+        return handleException(response);
+    }
+
+    /** enum type 일치하지 않아 binding 못할 경우 발생 주로 @RequestParam enum으로 binding 못했을 경우 발생 */
+    @ExceptionHandler(MethodArgumentTypeMismatchException.class)
+    protected ResponseEntity<ErrorResponse> handleMethodArgumentTypeMismatchException(
+            MethodArgumentTypeMismatchException e) {
+        log.error("handleMethodArgumentTypeMismatchException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.INVALID_PARAMETER);
+        return handleException(response);
+    }
+
+    /** 지원하지 않은 HTTP method 호출 할 경우 발생 */
+    @ExceptionHandler(HttpRequestMethodNotSupportedException.class)
+    protected ResponseEntity<ErrorResponse> handleHttpRequestMethodNotSupportedException(
+            HttpRequestMethodNotSupportedException e) {
+        log.error("handleHttpRequestMethodNotSupportedException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.METHOD_NOT_ALLOWED);
+        return handleException(response);
+    }
+
+    /** Authentication 객체가 필요한 권한을 보유하지 않은 경우 발생합 */
+    @ExceptionHandler(AccessDeniedException.class)
+    protected ResponseEntity<ErrorResponse> handleAccessDeniedException(AccessDeniedException e) {
+        log.error("handleAccessDeniedException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.HANDLE_ACCESS_DENIED);
+        return handleException(response);
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    protected ResponseEntity<ErrorResponse> handleIllegalArgumentException(Exception e) {
+        log.error("handleIllegalArgumentException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.INVALID_PARAMETER);
+        return handleException(response);
+    }
+
+    @ExceptionHandler(IllegalStateException.class)
+    protected ResponseEntity<ErrorResponse> handleIllegalStateException(Exception e) {
+        log.error("handleIllegalStateException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.INVALID_PARAMETER);
+        return handleException(response);
+    }
+
+    @ExceptionHandler(BusinessException.class)
+    protected ResponseEntity<ErrorResponse> handleBusinessException(final BusinessException e) {
+        log.error("handleBusinessException", e);
+        final ErrorCode errorCode = e.getErrorCode();
+        final ErrorResponse response = ErrorResponse.of(errorCode);
+        return handleException(response);
+    }
+
+    @ExceptionHandler(Exception.class)
+    protected ResponseEntity<ErrorResponse> handleException(Exception e) {
+        log.error("handleException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.UNKNOWN_ERROR);
+        return handleException(response);
+    }
+}
