+++ src/backend/common-module/common-reactive/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    // Reactive
+    implementation 'org.springframework.boot:spring-boot-starter-webflux'
+    testImplementation 'io.projectreactor:reactor-test'
+}
