+++ src/backend/common-module/common/src/main/java/com/lalala/event/LalalaEvent.java
@@ -1,0 +1,5 @@
+package com.lalala.event;
+
+public interface LalalaEvent {
+    String getTopic();
+}
