+++ src/backend/common-module/common/src/main/java/com/lalala/event/SearchMusicRequestEvent.java
@@ -1,0 +1,10 @@
+package com.lalala.event;
+
+public record SearchMusicRequestEvent(
+        Long id, String title, Short playTime, Long artistId, String artist, Long albumId, String album)
+        implements LalalaEvent {
+    @Override
+    public String getTopic() {
+        return TopicConstant.SEARCH_MUSIC_REQUEST.value;
+    }
+}
