+++ src/backend/common-module/common/src/main/java/com/lalala/event/StreamingCompleteEvent.java
@@ -1,0 +1,8 @@
+package com.lalala.event;
+
+public record StreamingCompleteEvent(Long userId, Long musicId) implements LalalaEvent {
+    @Override
+    public String getTopic() {
+        return TopicConstant.STREAMING_COMPLETE.value;
+    }
+}
