+++ src/backend/common-module/common/src/main/java/com/lalala/event/TopicConstant.java
@@ -1,0 +1,12 @@
+package com.lalala.event;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public enum TopicConstant {
+    STREAMING_COMPLETE("streaming_complete"),
+    SEARCH_MUSIC_REQUEST("search_music_request"),
+    ;
+
+    final String value;
+}
