+++ src/backend/common-module/common/src/main/java/com/lalala/passport/HMACEncoder.java
