--- src/backend/common-module/src/main/java/com/lalala/dto/LalalaEvent.java
@@ -1,5 +1,0 @@
-package com.lalala.dto;
-
-public interface LalalaEvent {
-    String getTopic();
-}
