--- src/backend/common-module/src/main/java/com/lalala/dto/event/StreamingCompleteEvent.java
@@ -1,10 +1,0 @@
-package com.lalala.dto.event;
-
-import com.lalala.dto.LalalaEvent;
-
-public record StreamingCompleteEvent(Long userId, Long musicId) implements LalalaEvent {
-    @Override
-    public String getTopic() {
-        return TopicConstant.STREAMING_COMPLETE.value;
-    }
-}
