--- src/backend/common-module/src/main/java/com/lalala/dto/event/TopicConstant.java
@@ -1,12 +1,0 @@
-package com.lalala.dto.event;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public enum TopicConstant {
-    STREAMING_COMPLETE("streaming_complete"),
-    SEARCH_MUSIC_REQUEST("search_music_request"),
-    ;
-
-    final String value;
-}
